feat: parse polyjuice log, only take used_gas as failed tx fee
