+++ crates/generator/Cargo.toml
@@ -18,7 +18,7 @@ gw-config = { path = "../config" }
 gw-store = { path = "../store" }
 gw-traits = { path = "../traits" }
 gw-ckb-hardfork = { path = "../ckb-hardfork" }
-gw-dynamic-config = { path = "../dynamic-config"}
+gw-utils = { path = "../utils"}
 anyhow = "1.0"
 blake2b-rs = "0.2"
 ckb-vm = { version = "=0.20.0-rc5", features = ["detect-asm"] }
