+++ crates/utils/Cargo.toml
@@ -15,3 +15,4 @@ faster-hex = "0.4"
 ckb-crypto = "0.100.0"
 sha3 = "0.9.1"
 secp256k1 = "0.21"
+log = "0.4"
