+++ crates/utils/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod fee;
 pub mod genesis_info;
+pub mod script_log;
 pub mod since;
 pub mod transaction_skeleton;
 pub mod wallet;
